+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+- Improve support for async functions and mutable borrows in contracts.
+
 ## 0.6.4
 
 - Add effective contract type to invariant violation messages.